==> scripts/04.2-train-svm.ipynb <==
import os
import pandas as pd
from sklearn.model_selection import train_test_split

os.chdir("C:\\Users\\lucac\\Documents\\GitHub\\song-cluster")

df = pd.read_csv('data/analysis_data/classifier_data.csv')

# Filter the dataset to include only the specified genres
df = df[df['genre_top'].isin([
    'Classical', 'Electronic', 'Experimental', 'Folk', 
    'Hip-Hop', 'Instrumental', 'International', 'Jazz', 
    'Pop', 'Rock'])]

feature_cols = [c for c in df.columns if c.startswith(('mfcc_', 'centroid_', 'rolloff_', 'zcr_'))]
X = df[feature_cols]
y = df['genre_top']

# Single train/test split (80/20), stratified by genre
X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.2,
    stratify=y,
    random_state=42
)

# Scale data
from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled  = scaler.transform(X_test)

# Train SVM classifier
from sklearn.svm import SVC
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix, 

# 3.1 Instantiate & train
svm = SVC(C=1.0, 
          kernel='rbf', 
          degree=3, 
          gamma='scale', 
          coef0=0.0, 
          shrinking=True, 
          probability=False, 
          tol=0.001, 
          cache_size=200, 
          class_weight= 'balanced', # balance for infrequent classes
          verbose=False, 
          max_iter=-1, 
          decision_function_shape='ovr', 
          break_ties=False, 
          random_state=None
        )

svm.fit(X_train_scaled, y_train)

# 3.2 Predict & evaluate
y_pred = svm.predict(X_test_scaled)
print("Accuracy:", accuracy_score(y_test, y_pred))
report = classification_report(y_test, y_pred, digits=3)
print(report)

# save table 
df_report = pd.DataFrame(report)
df_report.to_csv('other/tables/svm.csv', index=False)



# save svm model for later testing
import joblib
joblib.dump(svm, 'models/svm_model.pkl')
