==> scripts/06-predict.ipynb <==
import pandas as pd
import joblib
import os
from sklearn.metrics import accuracy_score
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import classification_report, confusion_matrix
from sklearn.model_selection import train_test_split

os.chdir("C:\\Users\\lucac\\Documents\\GitHub\\song-cluster")

# Load dataset
df = pd.read_csv('data/analysis_data/musicgpt_features.csv')

X_test = df.drop(["artist_name", "genre_top"], axis=1).values
y_test_label = df['genre_top']
y_test_values = df['genre_top'].values

# import most accurate (XGBoost) model
model = joblib.load('models/xgboost_model.pkl')

# import encoding table
encoding_table = pd.read_csv('models/encoding_table.csv')

# convert to dictionary mapping encoded class → label
encoding_dict = dict(zip(encoding_table['class'], encoding_table['label']))

# Scale test values for XGBoost
scaler = StandardScaler()
X_test_scaled = scaler.fit_transform(X_test)

# Make predictions on the test set
y_pred_values = model.predict(X_test_scaled)

# Unencode the predicted labels
y_pred_label = pd.Series(y_pred_values).map(encoding_dict).values

# Make a table of actual genres and predicted genres
predictions_df = pd.DataFrame({'MusicGPT Genre': y_test_label, 'Predicted': y_pred_label})
print(predictions_df)
